pyproject.toml
[project]
name = "agentproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
[tool.pyright]
venvPath = "."
venv = ".venv"

